Refactoring using async / await

+++ server/server.js
@@ -71,38 +71,66 @@ app.get('/todos/:id', authenticate, (req, res) => {
 }) 
 
 // DELETE /todo/1234
-app.delete('/todos/:id', authenticate, (req, res) => {
-	let id = req.params.id;
 
-	// Validate ID, if not valid then return 404
-	if (!ObjectID.isValid(id)) {
-		return res.status(404).send("Sorry, your ID was not valid");
-	}
+// OLD - Promises and Callback chaining
+// app.delete('/todos/:id', authenticate, (req, res) => {
+// 	let id = req.params.id;
 
-	// If valid, remove todo by ID
-	// Success
-	Todo.findOneAndRemove({
-		_id: id,
-		_creator: req.user._id
-	}).then((todo) => {
+// 	// Validate ID, if not valid then return 404
+// 	if (!ObjectID.isValid(id)) {
+// 		return res.status(404).send("Sorry, your ID was not valid");
+// 	}
+
+// 	// If valid, remove todo by ID
+// 	// Success
+// 	Todo.findOneAndRemove({
+// 		_id: id,
+// 		_creator: req.user._id
+// 	}).then((todo) => {
+// 		// If no doc, send 404
+// 		if (!todo) {
+// 			return res.status(404).send("Your todo item does not exist!");
+// 		}
+
+// 		// If doc, send doc back with 200
+// 		res.send({todo});
+
+
+// 	}).catch((e) => {
+// 		// Error
+// 			// 400
+// 			// Empty body
+// 		res.status(400).send();
+// 	})
+
+
+// });
+
+
+// NEW - Async and Await
+app.delete('/todos/:id', authenticate, async (req, res) => {
+	try {
+		let id = req.params.id;
+		// Validate ID, if not valid then return 404
+		if (!ObjectID.isValid(id)) {
+			return res.status(404).send("Sorry, your ID was not valid");
+		}
+
+		// If valid, remove todo by ID
+		// Success
+		let todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id})
+
 		// If no doc, send 404
 		if (!todo) {
 			return res.status(404).send("Your todo item does not exist!");
 		}
-
 		// If doc, send doc back with 200
 		res.send({todo});
-
-
-	}).catch((e) => {
-		// Error
-			// 400
-			// Empty body
-		res.status(400).send();
-	})
-
-
-});
+		
+	} catch {
+		res.status(400).send();	
+	}
+})
 
 // PATCH /todo/1234
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -155,43 +183,83 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 // POST /users/
 // Most similar is to create a new route for todo
 // Pick off the properties, email and password (and gives body variable to pass in)
-app.post('/users', (req, res) => {
-	let body = _.pick(req.body, ['email', 'password']);
-	let user = new User(body);
 
-	user.save().then(() => {
-		return user.generateAuthToken();
-	}).then((token) => {
+// OLD - Promises and Callback chaining
+// app.post('/users', (req, res) => {
+// 	let body = _.pick(req.body, ['email', 'password']);
+// 	let user = new User(body);
+
+// 	user.save().then(() => {
+// 		return user.generateAuthToken();
+// 	}).then((token) => {
+// 		res.header('x-auth', token).send(user);
+// 	}).catch((e) => {
+// 		res.status(400).send(e);
+// 	});
+
+// });
+
+// NEW - Async and Await
+app.post('/users', async (req, res) => {
+	try {
+		const body = _.pick(req.body, ['email', 'password']);
+		const user = new User(body);
+		await user.save();
+		const token = await user.generateAuthToken();
 		res.header('x-auth', token).send(user);
-	}).catch((e) => {
+	} catch (e) {
 		res.status(400).send(e);
-	});
-
+	}
 });
 
 app.get('/users/me', authenticate, (req, res) => {
 	res.send(req.user);
 });
 
-app.post('/users/login', (req,res) => {
-	let body = _.pick(req.body, ['email', 'password']);
+// OLD - Promises and Callback chaining
+// app.post('/users/login', (req,res) => {
+// 	let body = _.pick(req.body, ['email', 'password']);
 
-	User.findByCredentials(body.email, body.password).then((user) => {
-		return user.generateAuthToken().then((token) => {
-			res.header('x-auth', token).send(user);
-		});
-	}).catch((e) => {
+// 	User.findByCredentials(body.email, body.password).then((user) => {
+// 		return user.generateAuthToken().then((token) => {
+// 			res.header('x-auth', token).send(user);
+// 		});
+// 	}).catch((e) => {
+// 		res.status(400).send();
+// 	})
+
+// });
+
+// NEW - Async and Await
+app.post('/users/login', async (req,res) => {
+	try {
+		const body = _.pick(req.body, ['email', 'password']);
+		const user = await User.findByCredentials(body.email, body.password);
+		const token = await user.generateAuthToken();
+		res.header('x-auth', token).send(user);
+	} catch(e) {
 		res.status(400).send();
-	})
-
+	}
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-	req.user.removeToken(req.token).then(() => {
+
+// OLD - Promises and Callback chaining
+// app.delete('/users/me/token', authenticate, (req, res) => {
+// 	req.user.removeToken(req.token).then(() => {
+// 		res.status(200).send();
+// 	}, () => {
+// 		res.status(400).send();
+// 	});
+// });
+
+// NEW - Async and Await
+app.delete('/users/me/token', authenticate, async (req, res) => {
+	try {
+		await req.user.removeToken(req.token);
 		res.status(200).send();
-	}, () => {
+	} catch(e) {
 		res.status(400).send();
-	});
+	}
 });
 
 app.listen(port, () => {
